brain/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.11
pydantic==2.10.4
apscheduler==3.10.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
